test: ユニットテストファイルを分割して整理
「追加」ボタンのテストを、別ファイルに分割した。
